% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{col2hex2col-package}
\alias{col2hex2col}
\alias{col2hex2col-package}
\title{col2hex2col: Convert Between Color Names and Hex Codes}
\description{
Provides simple functions to convert between R color names and hexadecimal color codes. The package includes bidirectional conversion functions that work with single colors or vectors of colors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AnttiRask/col2hex2col}
  \item Report bugs at \url{https://github.com/AnttiRask/col2hex2col/issues}
}

}
\author{
\strong{Maintainer}: Antti Rask \email{anttilennartrask@gmail.com}

}
\keyword{internal}
